Рефакторинг кода  второй части задания отрисовки
