+++ js/main.js
@@ -1,7 +1,5 @@
-// import {pictures} from './rendering-data.js';
-// import {bigPictureImg} from './rendering-data.js';
 import { mocks } from './mock-data.js';
-import { renderPictureCards } from './rendering-data.js';
-renderPictureCards(mocks);
+import {renderPictures} from './render.js';
+renderPictures(mocks);
 // window.console.log(renderPictureCards(mocks,getPictureCardTemplate));
